Fix problem of RL78 GNURL78 port layer regarding code pointer size and potentially same problem of Renesas RL78 CC-RL port layer

+++ portable/GCC/RL78/port.c
@@ -105,7 +105,7 @@ uint32_t *pulLocal;
 
 	/* Task function start address combined with the PSW. */
 	pulLocal = ( uint32_t * ) pxTopOfStack;
-	*pulLocal = ( ( ( uint32_t ) pxCode ) | ( portPSW << 24UL ) );
+	*pulLocal = ( ( ( uint32_t ) ( uint16_t ) pxCode ) | ( portPSW << 24UL ) );
 	pxTopOfStack--;
 
 	/* An initial value for the AX register. */
